updating test

+++ circuit/test/test_phase.py
@@ -3,9 +3,11 @@ from circuit.phase import run_phase
 import os, os.path
 
 def test_run_phase():
-
-    os.remove('./circuit/figures/phase1.png')
-    os.remove('./circuit/figures/phase2.png')
+    
+    if os.path.isfile('./circuit/figures/phase1.png'):
+        os.remove('./circuit/figures/phase1.png')
+    if os.path.isfile('./circuit/figures/phase2.png'):
+        os.remove('./circuit/figures/phase2.png')
 
     run_phase()
